+++ packages/pds/src/api/com/atproto/server/createSession.ts
@@ -3,6 +3,7 @@ import { AuthRequiredError } from '@atproto/xrpc-server'
 import AppContext from '../../../../context'
 import { softDeleted } from '../../../../db/util'
 import { Server } from '../../../../lexicon'
+import { didDocForSession } from './util'
 
 export default function (server: Server, ctx: AppContext) {
   server.com.atproto.server.createSession({
@@ -54,16 +55,20 @@ export default function (server: Server, ctx: AppContext) {
         )
       }
 
-      const { access, refresh } = await authService.createSession({
-        did: user.did,
-        pdsDid: user.pdsDid,
-        appPasswordName,
-      })
+      const [{ access, refresh }, didDoc] = await Promise.all([
+        authService.createSession({
+          did: user.did,
+          pdsDid: user.pdsDid,
+          appPasswordName,
+        }),
+        didDocForSession(ctx, user.did),
+      ])
 
       return {
         encoding: 'application/json',
         body: {
           did: user.did,
+          didDoc,
           handle: user.handle,
           email: user.email,
           emailConfirmed: !!user.emailConfirmedAt,