+++ packages/pds/src/api/com/atproto/server/refreshSession.ts
@@ -2,6 +2,7 @@ import { AuthRequiredError, InvalidRequestError } from '@atproto/xrpc-server'
 import AppContext from '../../../../context'
 import { softDeleted } from '../../../../db/util'
 import { Server } from '../../../../lexicon'
+import { didDocForSession } from './util'
 
 export default function (server: Server, ctx: AppContext) {
   server.com.atproto.server.refreshSession({
@@ -21,13 +22,16 @@ export default function (server: Server, ctx: AppContext) {
         )
       }
 
-      const res = await ctx.db.transaction((dbTxn) => {
-        return ctx.services.auth(dbTxn).rotateRefreshToken({
-          id: auth.credentials.tokenId,
-          pdsDid: user.pdsDid,
-        })
-      })
-      if (res === null) {
+      const [didDoc, rotated] = await Promise.all([
+        didDocForSession(ctx, user.did),
+        ctx.db.transaction((dbTxn) => {
+          return ctx.services.auth(dbTxn).rotateRefreshToken({
+            id: auth.credentials.tokenId,
+            pdsDid: user.pdsDid,
+          })
+        }),
+      ])
+      if (rotated === null) {
         throw new InvalidRequestError('Token has been revoked', 'ExpiredToken')
       }
 
@@ -35,9 +39,10 @@ export default function (server: Server, ctx: AppContext) {
         encoding: 'application/json',
         body: {
           did: user.did,
+          didDoc,
           handle: user.handle,
-          accessJwt: res.access,
-          refreshJwt: res.refresh,
+          accessJwt: rotated.access,
+          refreshJwt: rotated.refresh,
         },
       }
     },