+++ packages/pds/src/lexicon/types/com/atproto/server/createAccount.ts
@@ -26,6 +26,7 @@ export interface OutputSchema {
   refreshJwt: string
   handle: string
   did: string
+  didDoc?: {}
   [k: string]: unknown
 }