+++ packages/repo/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atproto/repo
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [[`bb039d8e`](https://github.com/bluesky-social/atproto/commit/bb039d8e4ce5b7f70c4f3e86d1327e210ef24dc3), [`35d108ce`](https://github.com/bluesky-social/atproto/commit/35d108ce94866ce1b3d147cd0620a0ba1c4ebcd7)]:
+  - @atproto/identity@0.3.0
+  - @atproto/common-web@0.2.2
+  - @atproto/common@0.3.2
+  - @atproto/lexicon@0.2.3
+  - @atproto/syntax@0.1.3
+
 ## 0.3.2
 
 ### Patch Changes